==> 05_Budowa_schematu_reporting.ipynb <==
 # Opis notatnika
 Analiza przygotowana w poprzednim kroku została odebrana pozytywnie. W związku z tym, zostaliśmy poproszeni
 o przygotowanie raportu na podstawie danych z roku 2020. Ma być on później  wdrożony przez zespół IT na zewnętrzny
 serwis. Wybór padł na `Dash`.

 Zadanie wykonamy w dwóch krokach:
 1. Znając wymagania raportu, stworzymy na bazie danych odpowiednie komponenty, które zostaną następnie wykorzystane do wizualizacji.
 1. Stworzymy raport używając `Dash`.

 Ten notatnik skupia się na odpowiednim przygotowaniu bazy danych pod cele raportowe.

 # Wymagania biznesowe
 Po prezentacji wyników analizy na niezależnym spotkaniu, zostały określone następujące obszary raportowania:
 1. Wyświetlanie TOP 10 lotnisk na podstawie liczby odlotów, wyświetlona ma być również informacja o liczbie przylotów - widok `top_airports_by_departure`.
 1. Wyświetlenie TOP 10 tras na podstawie niezawodności, minimalna liczba lotów odbytych na trasie powinna wynosić co najmniej 10000. Przy czym istotna dla nas jest kolejność, przykładowo trasa (Warszawa, Paryż) jest inna niż (Paryż, Warszawa) - widok `top_reliability_roads`.
 1. Porównanie roku 2019 vs. 2020, aby śledzić wpływ COVID na realizację lotów. Interesują nas podejścia:
     - miesiąc do miesiąca, przykładowo styczeń 2019 vs. styczeń 2020, luty 2019 vs. luty 2020 itd. - widok `year_to_year_comparision`,
     - dzień do dnia, przykładowo wtorek 2019 vs. wtorek 2020 - widok `day_to_day_comparision`.
 1. Dzienny.

 # Podejście techniczne do problemu
 Naszym celem będzie odseparowanie warstwy przygotowania danych (logika raportu) od warstwy prezentacyjnej (wizualizacja).
 Chcemy zapewnić, aby odpowiednie procesy zajmowały się tylko swoimi zadaniami. W ten sposób warstwa prezentacyjna powinna w jak najmniejszym stopniu przetwarzać otrzymane dane.

 To podejście będzie spójne ze współczesnym sposobem projektowania aplikacji. Ma to też dodatkowy benefit - ze względu na wolumen danych na bazie, nie musimy ich najpierw pobierać - warstwa logiczna je odpowiednio zagreguje i przekaże zdecydowanie mniejszą liczbę wierszy, co przyśpieszy działanie całości.

 > Logika działania tego notebooka jest zbieżna z tą, którą robiliśmy już na przykładzie `Inicjowania bazy danych` - warto mieć go pod ręką.

 # Przygotowanie bazy danych
 Na bazie danych, gdzie umieszczone są już dane, wszystko wylądowało na schemacie `public`. Ponieważ mamy już dedykowany obszar wykorzystania danych, stworzymy sobie schemat dedykowany - `reporting`.
 Dalej stworzymy widoki, które odpowiedzą na zadane wcześniej pytania

 ## Stworzenie dedykowanego schematu
 W pliku `reporting.sql` napisz kwerendę, która stworzy (o ile już nie istnieje) schemat `reporting`.
 Ten temat nie był omawiany w trakcie trwania kursu, jednak łatwo można uzupełnić wiedzę czytając np. [ten](https://www.postgresqltutorial.com/postgresql-administration/postgresql-create-schema/) samouczek.

 # Aktualizacja bazy danych
 W tym miejscu odpowiednio skonfiguruj połączenie do bazy danych.

 Tutaj zaimportuj potrzebne biblioteki

from psycopg2 import connect

 ## Konfiguracja połączenia
 Tutaj uzupełnij konfigurację połączenia

username = "postgres"
passwd = "zaq1@WSX"
hostname = "localhost"  
db_name = "airlines"

 Tutaj zdefiniuj zmienną `con` oraz `cursor`

cnx = connect(user=username, password=passwd, host=hostname, database=db_name)

cursor = cnx.cursor()

 ## Wczytanie pliku `reporting.sql`
 Z katalogu `sql` wczytaj plik `reporting.sql`

--CREATE SCHEMA IF NOT EXIST reporting;

CREATE OR REPLACE VIEW reporting.flight as
SELECT *, 
 CASE
  WHEN dep_delay_new > 0 THEN 1
  ELSE 0
 END
 as is_delayed FROM public.flight WHERE cancelled != 0;
;

DROP VIEW reporting.top_reliability_roads;
CREATE OR REPLACE VIEW reporting.top_reliability_roads AS
SELECT 
    flight.origin_airport_id,
    airports.display_airport_name as origin_airport_name, 
    flight.dest_airport_id,
    airports_dest.display_airport_name as dest_airport_name,
    flight.year,
    count(*) as cnt,
    (cast(sum(CASE WHEN flight.dep_delay_new > 0 THEN 1 ELSE 0 END) as float) / count(*)) as reliability, 
    DENSE_RANK() OVER (ORDER BY (cast(sum(CASE WHEN flight.dep_delay_new > 0 THEN 1 ELSE 0 END) as float) / count(*)) ASC) as nb
FROM public.flight as flight JOIN public.airport_list as airports
                                ON flight.origin_airport_id = airports.origin_airport_id
                             JOIN public.airport_list as airports_dest
                                ON flight.dest_airport_id = airports_dest.origin_airport_id
GROUP BY (flight.year, flight.origin_airport_id, airports.display_airport_name, flight.dest_airport_id, airports_dest.display_airport_name)
HAVING count(*) > 10000
LIMIT 10
;

DROP VIEW reporting.year_to_year_comparision;
CREATE OR REPLACE VIEW reporting.year_to_year_comparision AS
SELECT
    flight.year,
    flight.month, 
    count(*) as flights_amount, 
    (cast(sum(CASE WHEN flight.dep_delay_new > 0 THEN 1 ELSE 0 END) as float) / count(*)) as reliability
FROM public.flight as flight
GROUP BY flight.year, flight.month
;

DROP VIEW reporting.day_to_day_comparision;
CREATE OR REPLACE VIEW reporting.day_to_day_comparision AS
SELECT
    flight.year,
    flight.day_of_week, 
    count(*) as flights_amount
FROM public.flight as flight
GROUP BY flight.year, flight.day_of_week
;

DROP VIEW reporting.day_by_day_reliability;
CREATE OR REPLACE VIEW reporting.day_by_day_reliability AS
SELECT
    to_date((cast(flight.year as text) || '-' || LPAD(cast(flight.month as text), 2, '0') || '-' || LPAD(cast(flight.day_of_month as text), 2, '0')), 'YYYY-MM-DD') as "date",
    (cast(sum(CASE WHEN flight.dep_delay_new > 0 THEN 1 ELSE 0 END) as float) / count(*)) as reliability
FROM public.flight as flight
GROUP BY "date"
;




 W tym miejscu odpowiednio rozdziel zawartość pliku `reporting.sql` na mniejsze kwerendy używając `;`

 W tym miejscu wykonaj każdą z kwerend, aby zainicjować strukturę bazy danych

 Zatwierdzenie wszystkich operacji na bazie, czyli stworzenie widoków

 ### Sprawdzenie

# Ten kod chyba wygląda znajomo....
# istnienie widoków możemy sprawdzić tak samo jak tabele
def check_if_table_exists(table_name):
    msg = f"Sprawdzam, czy istnieje tabela {table_name}"
    print(msg)

    query = f"select 1 from {table_name}"
    # jeżeli tabela nie istnieje, ten krok zwróci wyjątek
    cursor.execute(query)
    print('OK!')

views_to_test = [
    'reporting.flight',
    'reporting.top_reliability_roads',
    'reporting.year_to_year_comparision',
    'reporting.day_to_day_comparision',
    'reporting.day_by_day_reliability'
]

for view in views_to_test:
    check_if_table_exists(view)

cnx.close()
msg = "Wszystko wygląda OK :) Możesz przejść do kolejnego zadania."
print(msg)

 # Podsumowanie
 W tym notatniku stworzyliśmy nowy schemat - `reporting`, którego zadaniem jest przygotowanie naszych danych
 do wizualizacji. Dalsza część pracy będzie polegała na wyświetleniu w wizualnie atrakcyjny sposób danych w interaktywnym raporcie stworzonym
 za pomocą `Dash`.
